+++ .gitignore
@@ -3,3 +3,4 @@
 # Buidler
 **/cache/**
 **/artifacts/**
+.env
