+++ packages/api3-token/.env.example
@@ -1,0 +1,3 @@
+MAINNET_PROVIDER_URL=
+RINKEBY_PROVIDER_URL=
+MNEMONIC=
