+++ packages/api3-token/.gitignore
@@ -1,0 +1,1 @@
+FlatApi3Token.sol
