+++ packages/api3-token/buidler.config.js
@@ -1,5 +1,19 @@
+require('dotenv').config();
 usePlugin("@nomiclabs/buidler-waffle");
+usePlugin("@nomiclabs/buidler-etherscan");
 module.exports = {
+  networks: {
+    buidlerevm: {
+    },
+    mainnet: {
+      url: process.env.MAINNET_PROVIDER_URL,
+      accounts: {mnemonic: process.env.MNEMONIC}
+    },
+    rinkeby: {
+      url: process.env.RINKEBY_PROVIDER_URL,
+      accounts: {mnemonic: process.env.MNEMONIC}
+    },
+  },
   solc: {
     version: "0.6.12",
   },
